moved DB interface and implementation into different locations
