+++ src/test/java/lt/andriaus/hangman/usecase/RamDatabaseTest.java
@@ -1,12 +1,10 @@
-package lt.andriaus.hangman.database;
+package lt.andriaus.hangman.usecase;
 
 import lt.andriaus.hangman.database.Database;
-import lt.andriaus.hangman.database.RamDatabase;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Map;
